==> 6_descriptive_maps.Rmd <==
---
title: 'Part 6: Descriptive Map'
author: "Maggie Li (ml4424)"
date: "10/12/2020"
output: html_document
---

```{r}
library(tidyverse)
library(leaflet)
library(htmltools)
library(mapview)
```

## Descriptive Map of all the AI Counties

### Define AI counties as an sp file, by joining county lists of all 3 classifications with a county shapefile
```{r}
#map all native counties; to get sense of location to compare to monitor coverage
# setwd("/Users/maggieli/OneDrive - cumc.columbia.edu/Native Air Pollution Paper/Total PM paper/Drafts/ajph_submission/ajph_final_docs/figures/map")
counties <- readOGR("Data/cb_2018_us_county_500k/cb_2018_us_county_500k.shp")
states = readOGR("Data/cb_2018_us_state_500k/cb_2018_us_state_500k.shp")
summary(counties)

# read in counties by different classifications from 2_define_vars.Rmd
census 
cluster 
reservation 

#census
census_counties <- census %>%
  mutate(County = str_replace(County, "46113", "46102")) %>% 
  dplyr::rename(GEOID = County) 
census_counties$State <- substr(census_counties$GEOID, 0, 2)
census_counties$State <- as.numeric(census_counties$State)
census_counties <- census_counties %>% filter(State <= 56,
                                              State != 2,
                                              State != 15)
census_counties
census_shp <- merge(counties, census_counties, by = 'GEOID', all = FALSE, duplicateGeoms = TRUE)
length(census_counties$GEOID)
length(census_shp$GEOID) #excludes 1 county

#cluster
cluster_counties <- cluster %>%
  mutate(County = str_replace(County, "46113", "46102")) %>% 
  dplyr::rename(GEOID = County)
cluster_counties$State <- substr(cluster_counties$GEOID, 0, 2)
cluster_counties$State <- as.numeric(as.character(cluster_counties$State))
cluster_counties <- cluster_counties %>% filter(State <= 56,
                                              State != 2,
                                              State != 15)
cluster_counties
cluster_shp <- merge(counties, cluster_counties, by = 'GEOID', 
                     all = FALSE, duplicateGeoms = TRUE)
summary(cluster_shp)
length(cluster_counties$GEOID)
length(cluster_shp$GEOID) 

#reservation
reservation_counties <- reservation %>% 
  mutate(County = str_replace(County, "46113", "46102")) %>% 
  dplyr::rename(GEOID = County)
reservation_counties$State <- substr(reservation_counties$GEOID, 0, 2)
reservation_counties$State <- as.numeric(reservation_counties$State)
reservation_counties <- reservation_counties %>% 
  filter(State <= 56,
         State != 2,
         State != 15) 
unique(reservation_counties$GEOID)
reservation_shp <- merge(counties, reservation_counties, by = 'GEOID', 
                 all = FALSE, duplicateGeoms = TRUE)
length(reservation_counties$GEOID)
length(reservation_shp$GEOID) 


```

### Show overlap of AI county definitions (data visualized in venn diagram)
```{r}
# Non-restricted for each definition (main analysis)
census_counties <- census_counties %>%
  dplyr::select(GEOID)
cluster_counties <- cluster_counties %>%
  dplyr::select(GEOID)
reservation_counties <- reservation_counties %>%
  dplyr::select(GEOID)

#three-way overlap
all_overlap <- cluster_counties %>% 
  inner_join(census_counties) %>% 
  inner_join(reservation_counties) %>% 
  mutate(county_type = "all_class")
all_overlap
nrow(all_overlap) #number of counties that fit all three definitions

#two-way overlap
cluster_census <- cluster_counties %>% 
  inner_join(census_counties) %>% 
  anti_join(all_overlap) %>% 
  mutate(county_type = "cluster_census")
nrow(cluster_census)

cluster_reservation <- cluster_counties %>%
  inner_join(reservation_counties) %>% 
  anti_join(all_overlap) %>% 
  mutate(county_type = "cluster_reservation")
nrow(cluster_reservation)

census_reservation <- census_counties %>%
  inner_join(reservation_counties) %>% 
  anti_join(all_overlap) %>% 
  mutate(county_type = "census_reservation")
nrow(census_reservation)

#exclusive to each definition
census_only <- census_counties %>%
  anti_join(census_reservation) %>%
  anti_join(cluster_census) %>%
  anti_join(all_overlap) %>% 
  mutate(county_type = "census_only")
nrow(census_only)

# cluster_only <- cluster_counties %>%
#   anti_join(cluster_census) %>%
#   anti_join(cluster_reservation) %>%
#   anti_join(all_overlap)
# nrow(cluster_only) ##NO COUNTIES

reservation_only <- reservation_counties %>%
  anti_join(cluster_reservation) %>%
  anti_join(census_reservation) %>%
  anti_join(all_overlap) %>% 
  mutate(county_type = "reservation_only")
nrow(reservation_only)

## combine all AI classifications into one df
ai_class_map_df = rbind(census_only,
      reservation_only,
      census_reservation,
      cluster_census,
      all_overlap) %>%
  mutate(county_type = factor(county_type, 
                              levels= c("census_only", "reservation_only", 
                                        "census_reservation", "cluster_census", 
                                        "all_class"))) %>% 
  mutate(GEOID = str_replace(GEOID, "46102", "46113")) # to match with mapping
  
ai_class_map_df
## don't need to run below, because we are just mapping AI-populated counties on top of a blank counties basemap below

## pull non-AI counties from 2_define_vars.Rmd
nai_class_map_df = all_general_counties %>%
  mutate(county_type = case_when(
    county_type == 0 ~ "non-AI")) %>%
  rename(GEOID = County)

## join all AI and non-AI classified counties into one df
class_map_df = rbind(ai_class_map_df,
                     nai_class_map_df)
class_map_df
```

### Convert overlap counties to shp and only census and reservation classifications

```{r}
# Cluster x Census
cluster_census_counties <- cluster_census
cluster_census_counties$State <- substr(cluster_census_counties$GEOID, 0, 2)
cluster_census_counties$State <- as.numeric(cluster_census_counties$State)
cluster_census_counties <- cluster_census_counties %>% filter(State <= 56,
                                              State != 2,
                                              State != 15)
cluster_census_counties
cluster_census_shp <- merge(counties, cluster_census_counties, 
                    by = 'GEOID', all = FALSE, duplicateGeoms = TRUE)
length(cluster_census_counties$GEOID)
length(cluster_census_shp$GEOID)

# Cluster x Reservation -- dont run b/c no counties
# cluster_reservation_counties <- cluster_reservation
# cluster_reservation_counties$State <- substr(cluster_reservation_counties$GEOID, 0, 2)
# cluster_reservation_counties$State <- as.numeric(cluster_reservation_counties$State)
# cluster_reservation_counties <- cluster_reservation_counties %>% filter(State <= 56,
#                                               State != 2,
#                                               State != 15)
# cluster_reservation_counties
# cluster_reservation_shp <- merge(counties, cluster_reservation_counties, 
#                     by = 'GEOID', all = FALSE, duplicateGeoms = TRUE)
# length(cluster_reservation_counties$GEOID)
# length(cluster_reservation_shp$GEOID)

# Census x Reservation
census_reservation_counties <- census_reservation
census_reservation_counties$State <- substr(census_reservation_counties$GEOID, 0, 2)
census_reservation_counties$State <- as.numeric(census_reservation_counties$State)
census_reservation_counties <- census_reservation_counties %>% filter(State <= 56,
                                              State != 2,
                                              State != 15)
census_reservation_counties
census_reservation_shp <- merge(counties, census_reservation_counties, 
                    by = 'GEOID', all = FALSE, duplicateGeoms = TRUE)
length(census_reservation_counties$GEOID)
length(census_reservation_shp$GEOID)

# All overlap
all_overlap_counties <- all_overlap
all_overlap_counties$State <- substr(all_overlap_counties$GEOID, 0, 2)
all_overlap_counties$State <- as.numeric(all_overlap_counties$State)
all_overlap_counties <- all_overlap_counties %>% filter(State <= 56,
                                              State != 2,
                                              State != 15)
all_overlap_counties
all_overlap_shp <- merge(counties, all_overlap_counties, 
                    by = 'GEOID', all = FALSE, duplicateGeoms = TRUE)
length(all_overlap_counties$GEOID)
length(all_overlap_shp$GEOID)

# Only census
census_only_counties <- census_only
census_only_counties$State <- substr(census_only_counties$GEOID, 0, 2)
census_only_counties$State <- as.numeric(census_only_counties$State)
census_only_counties <- census_only_counties %>% filter(State <= 56,
                                              State != 2,
                                              State != 15)
census_only_counties
census_only_shp <- merge(counties, census_only_counties, 
                    by = 'GEOID', all = FALSE, duplicateGeoms = TRUE)
length(census_only_counties$GEOID)
length(census_only_shp$GEOID)

# Tribal entity only
reservation_only_counties <- reservation_only
reservation_only_counties$State <- substr(reservation_only_counties$GEOID, 0, 2)
reservation_only_counties$State <- as.numeric(reservation_only_counties$State)
reservation_only_counties <- reservation_only_counties %>% filter(State <= 56,
                                              State != 2,
                                              State != 15)
reservation_only_counties
reservation_only_shp <- merge(counties, reservation_only_counties, 
                    by = 'GEOID', all = FALSE, duplicateGeoms = TRUE)
length(reservation_only_counties$GEOID)
length(reservation_only_shp$GEOID)

```

Resource to create map: https://rstudio-pubs-static.s3.amazonaws.com/285359_35d630ddbc554c62b9137c13f063ee0e.html

```{r making base map}
library(ggmap)
library(mapdata)

#name the built-in state dataset in map_data to 'states'
states <- map_data("state")

#name the built-in county dataset in mapdata to 'counties'
counties <- map_data("county")

#create key to have column with state abbreviation to join with our dataset
state_key <- data.frame(state.name, state.abb) 
names(state_key) <- c("region","state_abb")
state_key = state_key %>% mutate(region = tolower(region),
                                 state_abb = tolower(state_abb))
print(state_key)

#convert counties region column from full name to abbreviation and rename to region to match our dataset
counties = counties %>% 
  left_join(state_key) %>% 
  dplyr::select(-region) %>% 
  rename(region = state_abb) 
counties

#view(counties)

base <- ggplot(data = states, mapping = aes(x = long, y = lat, group = group)) +
  coord_fixed(1.3) + geom_polygon(color = "black", fill = "white")
base

#add outlines of the counties on which to our stage map
base + geom_polygon(data = counties, fill = NA, color = "gray") + geom_polygon(color = "black", fill = NA)

```


```{r}
fipsnames<-read_csv("Data/fipsnames.csv")

fipsnames =
  fipsnames %>%
  rename("county_fips" = "FIPS") %>%
  rename("region" = "State",
         "subregion" =  "Name") %>%
  mutate(region = str_to_lower(region),
         subregion = str_to_lower(subregion))

# column with fips codes including leading zero to match class_map_df
GEOID <- formatC(fipsnames$county_fips, width = 5, format = "d", flag = "0")
fipsnames = cbind(fipsnames, GEOID)
fipsnames

#making categories of median, 95th percentile

#join with class_map_df with fipsnames and manually include miami dade, broomfield

class_map_df_join = class_map_df %>% 
  left_join(fipsnames, by = c("GEOID")) %>%
  mutate(region=ifelse(GEOID=="12086","fl",region)) %>%
  mutate(subregion=ifelse(GEOID=="12086","miami-dade",subregion)) %>%
  mutate(region=ifelse(GEOID=="08014","co",region)) %>%
  mutate(subregion=ifelse(GEOID=="08014","broomfield",subregion)) 
  

county_ai_map_df = left_join(counties, class_map_df_join, 
                              by = c("subregion", "region"), copy=FALSE) 
county_ai_map_df


# fill in NA value counties (washington DC, lagrange, obrien, ste genevieve, yellowstone natl park, suffolk, hampton, newport news, norfolk, virginia beach) as non-AI for county_type
subset(county_ai_map_df,is.na(county_type))

county_ai_map_df = county_ai_map_df %>%
  replace_na(list(county_type = "non-AI"))
county_ai_map_df

subset(county_ai_map_df,is.na(GEOID))

#add a fill to our county map based on data
#also remove all of the axes and borders and add a Title
county_ai_map <- base +
  geom_polygon(data = county_ai_map_df, 
               aes(fill = county_type), 
               color = "darkgray", 
               size =.05) +
  geom_polygon(color = "black", fill = NA) +
  theme_bw() 
county_ai_map

#transform the scale of the fill gradient to create more color variation
#also customize the legend breaks and a legend title

county_ai_map2 <- county_ai_map + 
  labs(fill = "AI-populated counties by classification", 
       y="Latitude", 
       x="Longitude") + 
  scale_fill_manual(values = c("#56B4E9", "#CC79A7","#0072B2", "#009E73", "#E69F00", "#FFFFFF"), 
                    labels = c("Al/AN County Census Population >5% Only", 
                               "County Area >20% within Fed Recognized Tribal Entity Only",
                               "Both Census and Tribal Entity Classified Only", 
                               "Both Census and Rural Cluster Classified Only", 
                               "Fits All Classification Schemes",
                               "Non-AI populated County")) +
  ggthemes::theme_map() + 
  theme(legend.title = element_text(hjust = 0.5, face = "bold", size=9),
        legend.position = "right",
        axis.text.x = element_blank(), axis.text.y = element_blank(),
        axis.ticks = element_blank()) + 
  guides(shape = guide_legend(override.aes = list(size = 8)),
               color = guide_legend(override.aes = list(size = 7)),
               shape = guide_legend(override.aes = list(size = 5))) +
        theme(legend.title = element_text(size = 7),
              legend.text  = element_text(size = 7))
county_ai_map2

ggsave(filename="figures/ai_map.pdf", plot=county_ai_map2, width=10, height=8, units="in")

```
