==> 8_supplement.Rmd <==
---
title: "Part 8: Supplement"
author: "Maggie Li (ml4424)"
date: "8/10/2020"
output: html_document
---

```{r}
library(xtable)
library(tidyverse)
```


# 1. Supplementary Tables by Land Zone Type and Geographical Region

## 1.1 Tables showing the number of AI and non-AI counties in each NCHS rural/urban classification
```{r}
#rural/urban categories for main analysis
all_avgmon_full
main_tbl_nchs <- table(all_avgmon_full$county_type,
      all_avgmon_full$land_use)
main_tbl_nchs 

in_tbls <- list(census_avgmon_full, 
                cluster_avgmon_full,
                reservation_avgmon_full)

#rural/urban list for separate classifications 
out_tbls <- list(length(in_tbls))
xtbls <- list()
for (i in 1:length(in_tbls)){
  out_tbls[[i]] <- table(in_tbls[[i]]$county_type,
                         in_tbls[[i]]$land_use)
  xtbls[[i]] <- xtable(out_tbls[[i]])
}
out_tbls #R tables
xtbls #xtables for LaTeX
```

##1.2 EPA/Geographical Region
```{r}
#EPA region for all classifications
all_avgmon_full
main_tbl_epa <- table(all_avgmon_full$county_type,
      all_avgmon_full$EPA_region)
main_tbl_epa

out_EPA_tbls <- list(length(in_tbls))
EPA_xtbls <- list()
for (i in 1:length(in_tbls)){
  out_EPA_tbls[[i]] <- table(in_tbls[[i]]$county_type,
                         in_tbls[[i]]$EPA_region)
  EPA_xtbls[[i]] <- xtable(out_EPA_tbls[[i]])
}
out_EPA_tbls #R tables
EPA_xtbls #xtables for LaTeX
```

#2. Monitor Distribution 
(Most of this not included in final version of manuscript)

##2.0 Define functions
```{r}
#1. Function to do get avg mon by county:
avg_mon_sens <- function(data, county_type){
  monitors <- data #different from OG function; already rbinded data before running it thru fxn
  monitors$county_type <- county_type
  monitors <- subset(monitors, select = c(County, num_mon, county_type, year))
  #make long -> wide with spread, in order to have columns of each year's values, and then find the mean of these columns
  monitors <- spread(monitors, key = year, value = num_mon, fill = 0)
  monitors$avg_mon <- rowMeans(monitors[c(seq(3,21))])
  #only select the county, county_type for IDing in the bar plot, and avg # of monitor value columns
  monitors <- subset(monitors,select = c(County, county_type, avg_mon))
  return(monitors)}

#2. Frequency function to get proportion of average monitors per county across all years, using output from avg_mon:
mon_freq <- function(data, county_type){
  freq <- plyr::count(data$avg_mon) #tbl with second column that records frequencies of the diff avg_mon values
  freq$prop <- freq$freq/
    sum(freq$freq) #new proportions column
  #rename x to avg_mon
  freq <- dplyr::rename(freq, avg_mon = x)
  freq$county_type <- county_type
  return(freq)} #for native and non-native, can check this output to see if area under curve adds up to 1

#3. Function to combine frequencies of monitors for AI/non-AI counties and aggregate bins 
combine_freq <- function(native_freq, general_freq){
  all_mon <- rbind(general_freq, native_freq)
  all_mon <- all_mon %>% 
    mutate(bins = case_when(
          avg_mon == 0 ~ "0",
          avg_mon > 0 & avg_mon < 1 ~ "(0,1)", 
          avg_mon >= 1 & avg_mon < 2 ~ "[1,2)",
          avg_mon >= 2 ~ ">=2"))
  all_mon$bins <- as.factor(all_mon$bins)
  final_mon_tbl <- all_mon %>%
    group_by(bins, county_type) %>%
    dplyr::summarise(total_prop=sum(prop),
                     total_freq=sum(freq)) %>% #summarize percent monitor (prop) and total num counties (freq) values for the four bins 
    mutate(lvl_bins = factor(bins, levels = c("0","(0,1)", "[1,2)", ">=2"))) %>%
    arrange(lvl_bins, county_type)
  return(final_mon_tbl)
}  
```


##2.1 Rural Counties (Main Analysis subset)
###2.1.1 Define counties
```{r}
#rbind lists into one df to join with covariates data in order to subset rural counties
nativePM25_df <- do.call(rbind, all_native_PM25)
#august 10 edit: need to subset out 46102 because these are double counted 
nativePM25_df <- nativePM25_df %>% subset(County != "46102")

generalPM25_df <- do.call(rbind, all_general_PM25)

#join with covariates data and subset out rural counties
rural_nativePM25 <- inner_join(nativePM25_df, covariates, by = "County") %>%
  subset(as.numeric(as.character(land_use)) >= 5)
rural_generalPM25 <- inner_join(generalPM25_df, covariates, by = "County") %>%
  subset(as.numeric(as.character(land_use)) >= 5)
```

###2.1.2 Create data tables & make bar graphs
```{r}
ruralnative_PM25_monitors <- avg_mon_sens(rural_nativePM25, "AI-Populated County")
ruralnative_PM25_freq <- mon_freq(ruralnative_PM25_monitors, "AI-Populated County")

ruralgeneral_PM25_monitors <- avg_mon_sens(rural_generalPM25, "Non-AI Populated County")
ruralgeneral_PM25_freq <- mon_freq(ruralgeneral_PM25_monitors, "Non-AI Populated County")

rural_PM25_tbl <- combine_freq(ruralnative_PM25_freq, ruralgeneral_PM25_freq)
sum(ruralgeneral_PM25_freq$prop)
sum(ruralnative_PM25_freq$prop)

#Bar graph all counties
rural_PM25_plot <- ggplot(rural_PM25_tbl, 
                            aes(x=lvl_bins, y = total_prop, fill=county_type)) + 
  theme_classic() + #define df cols for x and y axis, select minimal theme
  geom_bar(stat="identity", position = "dodge") +
  geom_text(aes(label=total_freq), position=position_dodge(width=0.9), vjust=-0.25) + #bin sample size above bars
  scale_y_continuous(labels = function(x) paste0(x*100, "%")) + #change y-axis to percent
  theme(plot.title = element_text(size = 20),
        axis.title = element_text(size = 14),
        axis.text = element_text(size = 12),
        legend.title = element_text(size = 14),
        legend.text = element_text(size = 12),
        legend.position = c(0.8, 0.8)) +
  labs(#title = expression(paste("Rural ", PM[2.5], " Monitor Distribution, 2000-2018")),
       x = "Average number of monitors per year by county type",
       y = "Percent of Total Counties",
       fill = "County Type") +
  scale_fill_manual("County Type", values = c("AI-Populated County" = "#5ab4ac", "Non-AI Populated County" = "#d8b365"))

rural_PM25_tbl
rural_PM25_plot 
ggsave("figures/supplement/rural_monitor_bar.png")

```

##2.2 Non-NE Counties (Main Analysis subset)
###2.2.1 Define counties
```{r}
#join with covariates data and subset out northeastern counties (EPA regions 1,2,3)
non_ne_nativePM25 <- inner_join(nativePM25_df, covariates, by = "County") %>%
  subset(as.numeric(as.character(EPA_region)) >= 4)
non_ne_generalPM25 <- inner_join(generalPM25_df, covariates, by = "County") %>%
  subset(as.numeric(as.character(EPA_region)) >= 4)
```

###2.2.2 Create data tables to make bar graphs
```{r}
non_ne_native_PM25_monitors <- avg_mon_sens(non_ne_nativePM25, "AI-Populated County")
non_ne_native_PM25_freq <- mon_freq(non_ne_native_PM25_monitors, "AI-Populated County")

non_ne_general_PM25_monitors <- avg_mon_sens(non_ne_generalPM25, "Non-AI Populated County")
non_ne_general_PM25_freq <- mon_freq(non_ne_general_PM25_monitors, "Non-AI Populated County")

non_ne_PM25_tbl <- combine_freq(non_ne_native_PM25_freq, non_ne_general_PM25_freq)
sum(non_ne_general_PM25_freq$prop)
sum(non_ne_native_PM25_freq$prop)

#Bar graph all counties
non_ne_PM25_plot <- ggplot(non_ne_PM25_tbl, 
                            aes(x=lvl_bins, y = total_prop, fill=county_type)) + 
  theme_classic() + #define df cols for x and y axis, select minimal theme
  geom_bar(stat="identity", position = "dodge") +
  geom_text(aes(label=total_freq), position=position_dodge(width=0.9), vjust=-0.25) + #bin sample size above bars
  scale_y_continuous(labels = function(x) paste0(x*100, "%")) + #change y-axis to percent
  theme(plot.title = element_text(size = 20),
        axis.title = element_text(size = 14),
        axis.text = element_text(size = 12),
        legend.title = element_text(size = 14),
        legend.text = element_text(size = 12),
        legend.position = c(0.8, 0.8)) +
  labs(#title = expression(paste("Non-NE ", PM[2.5], " Monitor Distribution, 2000-2018")),
       x = "Average number of monitors per year by county type",
       y = "Percent of Total Counties",
       fill = "County Type") +
  scale_fill_manual("County Type", values = c("AI-Populated County" = "#5ab4ac", "Non-AI Populated County" = "#d8b365"))

non_ne_PM25_tbl
non_ne_PM25_plot 
ggsave("figures/supplement/non_ne_monitor_bar.png")
```

##2.3 Main Analysis Counties + counties with any federally-recognized tribal reservations
###2.3.1 First define new tribal counties
```{r}
# QGIS: imported the tribal48 shapefile and the county shapefile, used the overlap function to join the layers (input layer = counties), which spits out a temporary layer showing the percent overlap of reservations in each county, for each county in the attribute table (I believe the column is called overlap_pc). I went into the properties of this layer, filtered out those with overlap_pc >= 0%, and exported this as a shapefile. I then read this in below:

# Read in cleaned counties with ANY area covered by reservation lands 
tribal_sensitivity <- rgdal::readOGR("Data/any_tribal_overlap.shp")

#map
leaflet() %>%
  addProviderTiles("CartoDB.PositronNoLabels") %>%
  addPolygons(data = tribal_sensitivity, color = "#3DACF0")

# change reservation dataset to be the same format
# first convert spdf to df
tribal_sensitivity <- as.data.frame(tribal_sensitivity)
tribal_sensitivity <- tribal_sensitivity %>% dplyr::rename(County = GEOID)
tribal_sensitivity$County <- as.character(tribal_sensitivity$County)

# need to recode FIPS 46102 to 46113 (Shannon County, SD was renamed Oglala Lakota County in 2015) for consistency with other data

tribal_sensitivity$County <- gsub("46102", "46113", tribal_sensitivity$County)

```

###2.3.2 Create new dataset of all native counties
```{r}
# Select common county column from all three classification definitions and join into single df
census <- native_census_counties %>% dplyr::select(County)
all_reservation <- tribal_sensitivity %>% dplyr::select(County)
cluster <- native_rural_counties %>% dplyr::select(County)

all_native_counties_sens <- rbind(census, all_reservation, cluster)

# only keep unique values, because classification definitions overlap
all_native_counties_sens <- unique(all_native_counties_sens) 

# County list of all non-native counties
all_counties_list <- all_counties %>% dplyr::select(GEOID) %>% 
  dplyr::rename(County = GEOID) #cleaned all_counties df to use in anti_join
all_general_counties_sens <- anti_join(x = all_counties_list,
                                  y = all_native_counties_sens,
                                  by= "County") %>% 
  filter(!str_detect(County, "^02"),
         !str_detect(County, "^15")) 

all_native_counties_sens$county_type <- 1
all_general_counties_sens$county_type <- 0

main_analysis_counties_sens <- rbind(all_native_counties_sens, all_general_counties_sens) %>% 
  mutate(County = str_replace(County, "46113", "46102"))
```

###2.3.3 Extract monitor PM2.5 concentrations and coverage:
```{r}
# Extract PM2.5 county monitor concentrations
all_native_PM25_sens <- list()
all_native_PM25_sens <- PM25_extract(PM25_all,
                                 all_native_counties_sens,
                                 all_native_PM25_sens) 
# Join all non-native counties with PM2.5 data
all_general_PM25_sens <- list()
all_general_PM25_sens <- PM25_extract(PM25_all,
                                 all_general_counties_sens,
                                 all_general_PM25_sens)
all_general_PM25_sens
all_native_PM25_sens
```

###2.3.4 Monitor coverage for all AI/non-AI counties (include main analysis counties and counties with any federally recognized reservation lands)
```{r}
native_PM25_monitors_sens <- avg_mon(all_native_PM25_sens, "AI-Populated County")
native_PM25_monitors_sens
native_PM25_freq_sens <- mon_freq(native_PM25_monitors_sens, "AI-Populated County")
general_PM25_monitors_sens <- avg_mon(all_general_PM25_sens, "Non-AI Populated County")
general_PM25_freq_sens <- mon_freq(general_PM25_monitors_sens, "Non-AI Populated County")

all_PM25_tbl_sens <- combine_freq(native_PM25_freq_sens, general_PM25_freq_sens)
# all_PM25_tbl$total_prop <- all_PM25_tbl$total_prop*100 #convert decimal-percentage to actual percentage
#check area under curve equals 1 for non-native and native
sum(general_PM25_freq_sens$prop)
sum(native_PM25_freq_sens$prop)

#Bar graph all counties
all_PM25_plot_sens <- ggplot(all_PM25_tbl_sens, 
                            aes(x=lvl_bins, y = total_prop, fill=county_type)) + 
  theme_classic() + #define df cols for x and y axis, select minimal theme
  geom_bar(stat="identity", position = "dodge") +
  geom_text(aes(label=total_freq), position=position_dodge(width=0.9), vjust=-0.25) + #bin sample size above bars
  scale_y_continuous(labels = function(x) paste0(x*100, "%")) + #change y-axis to percent
  theme(plot.title = element_text(size = 20),
        axis.title = element_text(size = 14),
        axis.text = element_text(size = 12),
        legend.title = element_text(size = 14),
        legend.text = element_text(size = 12),
        legend.position = c(0.8, 0.8)) +
  labs(#title = expression(paste(PM[2.5], " Monitor Distribution, 2000-2018 (Any Tribal Lands)")),
       x = "Average number of monitors per year by county type",
       y = "Percent of Total Counties",
       fill = "County Type") +
  scale_fill_manual("County Type", values = c("AI-Populated County" = "#5ab4ac", "Non-AI Populated County" = "#d8b365"))

all_PM25_tbl_sens
all_PM25_plot_sens 
ggsave("figures/supplement/any_tribal_monitor_bar.png")
```


##2.4 AI populated counties (census 5% definition only)

###2.4.1 Monitor coverage bar plot
```{r}
#CENSUS avg monitors
native_PM25_census_monitors <- avg_mon(native_PM25_census, "AI-Populated County")
native_PM25_census_freq <- mon_freq(native_PM25_census_monitors, "AI-Populated County")
general_PM25_census_monitors <- avg_mon(general_PM25$census, "Non-AI County")
general_PM25_census_freq <- mon_freq(general_PM25_census_monitors, "Non-AI Populated County")
census_PM25_tbl <- combine_freq(native_PM25_census_freq, general_PM25_census_freq)
census_PM25_tbl 
native_PM25_census_freq

#check area under curve equals 1 for non-native and native
sum(general_PM25_census_freq$prop)
sum(native_PM25_census_freq$prop)

#Bar graph census counties
census_PM25_plot <- ggplot(census_PM25_tbl, 
                            aes(x=lvl_bins, 
                                y = total_prop,
                                fill=county_type)) + 
  theme_classic() + #define df cols for x and y axis, select minimal theme
  geom_bar(stat="identity", position = "dodge") +
  geom_text(aes(label=total_freq), position=position_dodge(width=0.9), vjust=-0.25) + #bin sample size above bars
  scale_y_continuous(labels = function(x) paste0(x*100, "%")) + #change y-axis to percent
  theme(plot.title = element_text(size = 20),
        axis.title = element_text(size = 14),
        axis.text = element_text(size = 12),
        legend.title = element_text(size = 14),
        legend.text = element_text(size = 12),
        legend.position = c(0.8, 0.8)) +
  labs(#title = expression(paste(PM[2.5], " Monitor Distribution \n in AI vs Non-AI Populated Census Counties, 2000-2018")),
       x = "Average number of monitors per year by county type",
       y = "Percent of Total Counties",
       fill = "County Type") +
  scale_fill_manual("County Type", values = c("AI-Populated County" = "#5ab4ac", "Non-AI Populated County" = "#d8b365"))
census_PM25_plot 
ggsave("figures/supplement/census_monitor_bar.png")
```

##2.5 AI populated counties (rural cluster definition only)

###2.5.1 Monitor coverage bar plot
```{r}
#CLUSTER avg annual mon
native_PM25_cluster_monitors <- avg_mon(native_PM25, "AI-Populated County")
native_PM25_cluster_freq <- mon_freq(native_PM25_cluster_monitors, "AI-Populated County")
general_PM25_cluster_monitors <- avg_mon(general_PM25$cluster, "Non-AI Populated County")
general_PM25_cluster_freq <- mon_freq(general_PM25_cluster_monitors, "Non-AI Populated County")

cluster_PM25_tbl <- combine_freq(general_PM25_cluster_freq, native_PM25_cluster_freq)
sum(general_PM25_cluster_freq$prop)
sum(native_PM25_cluster_freq$prop)

# need to add in "fake" extra row for >=2 natives (since 0 native counties have more than 2 monitors)
extra_row <- data.frame(">=2",
               "AI-Populated County",
               0.001,
               0,
               ">=2")
names(extra_row) <- c(names(cluster_PM25_tbl))
extra_row$county_type <- as.character(extra_row$county_type)
cluster_PM25_tbl
cluster_PM25_tbl <- bind_rows(cluster_PM25_tbl, extra_row)
cluster_PM25_tbl$lvl_bins <- as.factor(cluster_PM25_tbl$lvl_bins)
cluster_PM25_tbl
#reorder factors
cluster_PM25_tbl$lvl_bins <- factor(cluster_PM25_tbl$lvl_bins, 
                                    levels = c("0", "(0,1)", "[1,2)", ">=2"))

cluster_PM25_tbl
cluster_PM25_plot <- ggplot(cluster_PM25_tbl, 
                            aes(x=lvl_bins, 
                                y = total_prop,
                                fill=county_type)) + 
  theme_classic() + #define df cols for x and y axis, select minimal theme
  geom_bar(stat="identity", position = "dodge") +
  geom_text(aes(label=total_freq), position=position_dodge(width=0.9), vjust=-0.25) + #bin sample size above bars
  scale_y_continuous(labels = function(x) paste0(x*100, "%")) + #change y-axis to percent
  theme(plot.title = element_text(size = 20),
        axis.title = element_text(size = 14),
        axis.text = element_text(size = 12),
        legend.title = element_text(size = 14),
        legend.text = element_text(size = 12),
        legend.position = c(0.8, 0.8)) +
  labs(#title = expression(paste(PM[2.5], " Monitor Distribution in Rural Clustered Counties, 2000-2018")),
       x = "Average number of monitors per year by county type",
       y = "Percent of Total Counties",
       fill = "County Type") +
  scale_fill_manual("County Type", values = c("AI-Populated County" = "#5ab4ac", "Non-AI Populated County" = "#d8b365"))
cluster_PM25_plot
ggsave("figures/supplement/cluster_monitor_bar.png")
```

##2.6 AI populated counties (20% areal overlap with reservations definition only)

###2.6.1 Monitor coverage bar plot
```{r}
reservation_PM25_monitors <- avg_mon(reservation_PM25, "AI-Populated County")
reservation_PM25_freq <- mon_freq(reservation_PM25_monitors, "AI-Populated County")

general_PM25_reservations_monitors <-avg_mon(general_PM25$reservations, "Non-AI Populated County")
general_reservation_PM25_freq <- mon_freq(general_PM25_reservations_monitors,"Non-AI Populated County")

reservation_PM25_tbl <- combine_freq(reservation_PM25_freq,
                                     general_reservation_PM25_freq)
reservation_PM25_plot <- ggplot(reservation_PM25_tbl, 
                            aes(x=lvl_bins, 
                                y = total_prop,
                                fill=county_type)) + 
  theme_classic() + #define df cols for x and y axis, select minimal theme
  geom_bar(stat="identity", position = "dodge") +
  geom_text(aes(label=total_freq), position=position_dodge(width=0.9), vjust=-0.25) + #bin sample size above bars
  scale_y_continuous(labels = function(x) paste0(x*100, "%")) + #change y-axis to percent
  theme(plot.title = element_text(size = 20),
        axis.title = element_text(size = 14),
        axis.text = element_text(size = 12),
        legend.title = element_text(size = 14),
        legend.text = element_text(size = 12),
        legend.position = c(0.8, 0.8)) +
  labs(#title = expression(paste(PM[2.5], " Monitor Distribution (Reservations), 2000-2018")),
       x = "Average number of monitors per year by county type",
       y = "Percent of Total Counties",
       fill = "County Type") +
  scale_fill_manual("County Type", values = c("AI-Populated County" = "#5ab4ac", "Non-AI Populated County" = "#d8b365"))
sum(reservation_PM25_freq$prop)
sum(general_reservation_PM25_freq$prop)
reservation_PM25_plot
ggsave("figures/supplement/reservation_monitor_bar.png")

reservation_PM25_tbl
```

# 3. Linear Mixed Effects Regression Sensitivity Analyses (Output = Modeled PM2.5) 

## 3.1 Rural Counties (Main Analysis subset); commented out because this was already generated in 7_stat_analysis.Rmd

```{r}
# # define rural counties
# ruralPM25_counties <- main_analysis_counties %>% inner_join(covariates) %>% subset(as.numeric(as.character(land_use)) >=5)
# 
# # inner join with rural counties dataset to get modeled PM2.5 values for these counties
# rural_model_PM25 <- inner_join(model_PM25, covariates) %>% inner_join(ruralPM25_counties)
# rural_model_PM25
```

###3.1.2 Run lmer model
```{r}
# # Split population density and hhincome into deciles for model
# rural_model_PM25$popd_q <- cut(rural_model_PM25$pop_density, quantile(rural_model_PM25$pop_density, seq(0,1,0.1)), include.lowest = TRUE)
# rural_model_PM25$hhinc_q <- cut(rural_model_PM25$hh_income, quantile(rural_model_PM25$hh_income, seq(0,1,0.1)), include.lowest = TRUE)
# rural_model_PM25
# sum(table(rural_model_PM25$popd_q, exclude = NULL)) == dim(rural_model_PM25)[1]
# 
# # edit cols to fit model function for linear mixed effects model
# rural_model_PM25$county_type <- as.factor(as.character(rural_model_PM25$county_type))
# 
#   # set referent
# rural_model_PM25$county_type <- relevel(rural_model_PM25$county_type,
#                                          ref = "0")
# rural_model_PM25$Year <- as.factor(rural_model_PM25$Year)
# rural_model_PM25 <- dplyr::rename(rural_model_PM25, year=Year)
# 
# # rename column for PM25 to match predefined function arguments
# rural_model_lmer <-  dplyr::rename(rural_model_PM25,annual_mean_all=PM25)
# 
# # Run LMER (fully adjusted only)
# summary(detrend_lmer_full(rural_model_lmer))
# summary(detrend_lmer_full(rural_model))
# 
# # point estimate
# summary(detrend_lmer_full(rural_model_lmer))$coefficients[2,1]
# # 95% upper
# summary(detrend_lmer_full(rural_model_lmer))$coefficients[2,1] + 1.96*summary(detrend_lmer_full(rural_model_lmer))$coefficients[2,2]
# 
# # 95% lower
# summary(detrend_lmer_full(rural_model_lmer))$coefficients[2,1] - 1.96*summary(detrend_lmer_full(rural_model_lmer))$coefficients[2,2]
# 
# # tex table output
# stargazer(detrend_lmer_full(rural_model_lmer),ci=T)
```

##3.2 Non-NE Counties (Main Analysis subset)

###3.2.1 Define Non-NE counties and join with model PM2.5 data
```{r}
non_ne_PM25_counties <- main_analysis_counties %>% 
  inner_join(covariates) %>% 
  subset(as.numeric(as.character(EPA_region)) >=4) %>% 
  mutate(county_type=as.factor(county_type)) %>% 
  dplyr::select(County)

# inner join with rural counties dataset to get modeled PM2.5 values for these counties
non_ne_model_PM25 <- model_lmer %>% inner_join(non_ne_PM25_counties)
non_ne_model_PM25
```

###3.2.2 Run lmer model
```{r}
# Run LMER (fully adjusted only)
non_ne_model_lmer = summary(detrend_lmer_full(non_ne_model_PM25))

# point estimate
non_ne_model_lmer$coefficients[2,1]
# 95% upper
non_ne_model_lmer$coefficients[2,1] + 1.96*non_ne_model_lmer$coefficients[2,2]

# 95% lower
non_ne_model_lmer$coefficients[2,1] - 1.96*non_ne_model_lmer$coefficients[2,2]

# tex table output
stargazer(non_ne_model_lmer,ci=T)
```

##3.2.3 Re-read in Model PM2.5 data (need to run this chunk before everything below due to overwriting variable name due to new definitions of county type) 
```{r}
# Read in model data into list of length 19 (# study years)
model_PM25 <- list()
DC_model_PM25 <- list()
yr = 2000
for (i in 1:19){
  model_PM25[[i]] <- read_csv(file = paste('Data/model_PM25/model_PM25_', 
                                           yr, '.csv', sep = '')) %>% 
    dplyr::select(c(FIPS, PM25, State))
  DC_model_PM25[[i]] <- read_csv(file = paste('Data/model_PM25/DC_data/DC_modelPM25_', 
                                           yr, '.csv', sep = '')) %>% 
    dplyr::select(c(FIPS, PM25, State)) # join DC data to this!
  
  model_PM25[[i]] <- rbind(model_PM25[[i]], DC_model_PM25[[i]]) %>% 
    mutate(Year = yr)
  yr <- yr + 1
}
model_PM25
model_PM25 <- do.call(rbind, model_PM25)
model_PM25 <- model_PM25 %>% 
  dplyr::rename(County=FIPS, 
                annual_mean_all=PM25,
                year=Year) %>% 
  mutate(State = as.numeric(State),
         year = as.factor(year)) %>% 
  inner_join(covariates) 
model_PM25

# Split population density and hhincome into deciles for model
model_PM25$popd_q <- cut(model_PM25$pop_density, quantile(model_PM25$pop_density, seq(0,1,0.1)), include.lowest = TRUE)
model_PM25$hhinc_q <- cut(model_PM25$hh_income, quantile(model_PM25$hh_income, seq(0,1,0.1)), include.lowest = TRUE)
model_PM25
sum(table(model_PM25$popd_q, exclude = NULL)) == dim(model_PM25)[1]
```

##3.3 Main Analysis + counties with any federally-recognized tribal entity 

###3.3.1 Define counties and join with model PM2.5 data
```{r}
main_analysis_counties_sens

# inner join with rural counties dataset to get modeled PM2.5 values for these counties
any_tribal_model_PM25 <- model_PM25 %>% 
  inner_join(main_analysis_counties_sens) %>% 
  mutate(county_type = as.factor(as.character(county_type)),
         county_type = relevel(county_type, ref = "0"))

```

###3.3.2 Run lmer model
```{r}
# Run LMER (fully adjusted only)
any_tribal_model_lmer = summary(detrend_lmer_full(any_tribal_model_PM25))

# point estimate
any_tribal_model_lmer$coefficients[2,1]
# 95% upper
any_tribal_model_lmer$coefficients[2,1] + 1.96*any_tribal_model_lmer$coefficients[2,2]

# 95% lower
any_tribal_model_lmer$coefficients[2,1] - 1.96*any_tribal_model_lmer$coefficients[2,2]

# tex table output
stargazer(any_tribal_model_lmer,ci=T)
```

##3.4 Census Classification only

###3.4.1 Define counties and join with model PM2.5 data
```{r}
# relevant df for AI/non-AI counties
general_counties_census 
native_census_counties_list

# new df combining the two
census_counties_df <- general_counties_census %>% 
  mutate(county_type = 0) %>%
  bind_rows(native_census_counties_list %>% mutate(county_type = 1)) %>% 
  mutate(County = str_replace(County, "46113", "46102"))

# join this with covariates & model PM2.5
census_model_PM25 <- model_PM25 %>% 
  inner_join(census_counties_df) %>% 
  mutate(county_type = as.factor(as.character(county_type)),
         county_type = relevel(county_type, ref = "0"))
census_model_PM25
```

###3.4.2  Run lmer model
```{r}
# Split population density and hhincome into deciles for model
census_model_PM25$popd_q <- cut(census_model_PM25$pop_density, quantile(census_model_PM25$pop_density, seq(0,1,0.1)), include.lowest = TRUE)
census_model_PM25$hhinc_q <- cut(census_model_PM25$hh_income, quantile(census_model_PM25$hh_income, seq(0,1,0.1)), include.lowest = TRUE)
census_model_PM25
sum(table(census_model_PM25$popd_q, exclude = NULL)) == dim(census_model_PM25)[1]

# Run LMER (fully adjusted only)
census_model_lmer= summary(detrend_lmer_full(census_model_PM25))

# point estimate
census_model_lmer$coefficients[2,1]
# 95% upper
census_model_lmer$coefficients[2,1] + 1.96*census_model_lmer$coefficients[2,2]

# 95% lower
census_model_lmer$coefficients[2,1] - 1.96*census_model_lmer$coefficients[2,2]

# tex table output
stargazer(census_model_lmer,ci=T)
```

##3.5 Rural Cluster classification only

###3.5.1 Define counties and join with covariates and model PM2.5 data
```{r}
# relevant df for AI/non-AI counties
general_counties_cluster
native_rural_counties_list

# new df combining the two
cluster_counties_df <- general_counties_cluster %>% mutate(county_type = 0) %>%
  bind_rows(native_rural_counties_list %>% mutate(county_type = 1)) %>% 
  mutate(County = str_replace(County, "46113", "46102"))

# join this with covariates & model PM2.5
cluster_model_PM25 <- model_PM25 %>% inner_join(cluster_counties_df) %>% 
  mutate(county_type = as.factor(as.character(county_type)),
         county_type = relevel(county_type, ref = "0"))
cluster_model_PM25
```

###3.5.2 Run lmer model
```{r}
# Run LMER (fully adjusted only)
cluster_model_lmer = summary(detrend_lmer_full(cluster_model_PM25))

# point estimate
cluster_model_lmer$coefficients[2,1]
# 95% upper
cluster_model_lmer$coefficients[2,1] + 1.96*cluster_model_lmer$coefficients[2,2]

# 95% lower
cluster_model_lmer$coefficients[2,1] - 1.96*cluster_model_lmer$coefficients[2,2]

# tex table output
stargazer(cluster_model_lmer,ci=T)
```

##3.6 Reservations classification only
```{r}
# relevant df for AI/non-AI counties
general_counties_reservations
reservation_counties

# new df combining the two
reservation_counties_df <- general_counties_reservations %>% 
  mutate(county_type = 0) %>%
  bind_rows(reservation_counties %>% rename(County = GEOID) %>% 
              dplyr::select(County) %>% mutate(county_type = 1)) %>% 
  subset(County != "46113")

# join this with covariates & model PM2.5
reservation_model_PM25 <- model_PM25 %>% inner_join(reservation_counties_df) %>% 
  mutate(county_type = as.factor(as.character(county_type)),
         county_type = relevel(county_type, ref = "0"))
reservation_model_PM25
```

###3.6.2 run lmer model
```{r}
# Run LMER (fully adjusted only)
reservation_model_lmer = summary(detrend_lmer_full(reservation_model_PM25))

# point estimate
reservation_model_lmer$coefficients[2,1]
# 95% upper
reservation_model_lmer$coefficients[2,1] + 1.96*reservation_model_lmer$coefficients[2,2]

# 95% lower
reservation_model_lmer$coefficients[2,1] - 1.96*reservation_model_lmer$coefficients[2,2]

# tex table output
stargazer(reservation_model_lmer,ci=T)
```

##3.7 Reservations classification only (any tribal overlap)
```{r}
tribal_sensitivity %>% dplyr::select(County)
general_counties_tribal_sens <- anti_join(x = all_counties_list,
                                     y = tribal_sensitivity %>% dplyr::select(County),
                                     by= "County")
# new df combining the two
tribal_sens_counties_df <- general_counties_tribal_sens %>% mutate(county_type = 0) %>%
  bind_rows(tribal_sensitivity %>% dplyr::select(County) %>% mutate(county_type = 1)) %>% 
  mutate(County = str_replace(County, "46113", "46102")) # to match with model PM2.5 data
tribal_sens_counties_df

# join this with covariates & model PM2.5
tribal_sens_model_PM25 <- model_PM25 %>% inner_join(tribal_sens_counties_df) %>% 
  mutate(county_type = as.factor(as.character(county_type)),
         county_type = relevel(county_type, ref = "0"))
tribal_sens_model_PM25 
```


```{r}
# Run LMER (fully adjusted only)
tribal_sens_model_lmer = summary(detrend_lmer_full(tribal_sens_model_PM25))

# point estimate
tribal_sens_model_lmer$coefficients[2,1]
# 95% upper
tribal_sens_model_lmer$coefficients[2,1] + 1.96*tribal_sens_model_lmer$coefficients[2,2]

# 95% lower
tribal_sens_model_lmer$coefficients[2,1] - 1.96*tribal_sens_model_lmer$coefficients[2,2]

# tex table output
stargazer(tribal_sens_model_lmer,ci=T)
```
